+++ extensions/Renamer/src/Renamer.Tests/Events/AutoRenamerRevertLogBatchTests.cs
@@ -4,7 +4,6 @@ using Cove.Plugins;
 using Microsoft.EntityFrameworkCore;
 using Renamer.Execution;
 using Renamer.Options;
-using Renamer.Planner;
 using Renamer.Tests.Execution;
 using Renamer.Tests.TestSupport;