+++ extensions/Renamer/src/Renamer.Tests/Events/DetachedElevationTests.cs
@@ -7,7 +7,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Renamer.Execution;
 using Renamer.Jobs;
 using Renamer.Options;
-using Renamer.Planner;
 using Renamer.Tests.Execution;
 using Renamer.Tests.TestSupport;